Worked on another way to get token from header, created a user controller and updated its routes

+++ server/controller/entries.js
@@ -7,18 +7,25 @@ class EntriesController {
    */
   static fetchEntries(req, res) {
     // retrieve all entries from the database FOR NOW!!!.
-    const fetchEntryQuery = 'SELECT * FROM entries;';
-    pool.query(fetchEntryQuery, (err, result) => {
-      if (result.rowCount < 1) {
-        return res.status(404).json({
-          message: 'No entry',
+    const fetchEntryQuery = `SELECT * FROM entries WHERE userid = '${req.decoded.userId}';`;
+    try {
+      pool.query(fetchEntryQuery, (err, result) => {
+        if (result.rowCount < 1) {
+          return res.status(404).json({
+            message: 'No entry',
+          });
+        }
+        return res.status(200).json({
+          message: 'All entries by users',
+          entries: result.rows,
         });
-      }
-      return res.status(200).json({
-        message: 'All entries by users',
-        entries: result.rows,
       });
-    });
+    } catch (err) {
+      return res.status(500).json({
+        message: 'Something went wrong!',
+      });
+    }
+    return null;
   }
 
   /**
@@ -26,19 +33,26 @@ class EntriesController {
    */
   static fetchSingleEntry(req, res) {
     const { entriesId } = req.params;
+    const findEntryQuery = `SELECT * FROM entries WHERE entriesId = '${entriesId}' AND userid = '${req.decoded.userId}';`;
     // Searches the particular entry from the database
-    const findEntryQuery = `SELECT * FROM entries WHERE entriesId = '${entriesId}';`;
-    pool.query(findEntryQuery, (err, result) => {
-      if (result.rowCount < 1) {
-        return res.status(404).json({
-          message: 'No such entry',
+    try {
+      pool.query(findEntryQuery, (err, result) => {
+        if (result.rowCount < 1) {
+          return res.status(404).json({
+            message: 'No such entry',
+          });
+        }
+        return res.status(200).json({
+          message: 'Entry found',
+          entry: result.rows[0],
         });
-      }
-      return res.status(200).json({
-        message: 'Entry found',
-        entry: result.rows[0],
       });
-    });
+    } catch (err) {
+      return res.status(500).json({
+        message: 'Something went wrong!',
+      });
+    }
+    return null;
   }
 
   /**
@@ -49,7 +63,7 @@ class EntriesController {
   static createEntry(req, res) {
     // Get parameters from the req.body
     const {
-      entryTitle, entry, visibility, userId,
+      entryTitle, entry, visibility,
     } = req.body;
     // Validation happens here
     if (InputValidator.validateEntryTitle(entryTitle) === false) {
@@ -68,7 +82,7 @@ class EntriesController {
       });
     }
     // Gets userId from the token
-    const createEntryQuery = `INSERT INTO entries (entry, entryTitle, visibility, userId) VALUES ('${entry}', '${entryTitle}', '${visibility}', '${userId}');`;
+    const createEntryQuery = `INSERT INTO entries (entry, entryTitle, visibility, userId) VALUES ('${entry}', '${entryTitle}', '${visibility}', '${req.decoded.userId}');`;
     pool.query(createEntryQuery, (err, result) => {
       if (result.rowCount < 1) {
         return res.status(400).json({
@@ -111,7 +125,7 @@ class EntriesController {
     }
 
     // Update SQL query
-    const updateEntryQuery = `UPDATE entries SET entryTitle = '${entryTitle}', entry = '${entry}', visibility = '${visibility}' WHERE entriesId = '${parseInt(entriesId, 10)}';`;
+    const updateEntryQuery = `UPDATE entries SET entryTitle = '${entryTitle}', entry = '${entry}', visibility = '${visibility}' WHERE entriesId = '${parseInt(entriesId, 10)}' AND userId = '${req.decoded.userId}' RETURNING *;`;
     pool.query(updateEntryQuery, (err, result) => {
       if (result.rowCount < 0) {
         return res.status(404).json({
@@ -120,7 +134,7 @@ class EntriesController {
       }
       return res.status(200).json({
         message: 'successfully updated',
-        entry: result.rows,
+        entry: result.rows[0],
       });
     });
     return null;
@@ -134,7 +148,7 @@ class EntriesController {
     const { entriesId } = req.params;
 
     // Delete entry from entries where entryId is something
-    const deleteEntryQuery = `DELETE FROM entries WHERE entriesId = ${entriesId};`;
+    const deleteEntryQuery = `DELETE FROM entries WHERE entriesId = '${entriesId}' AND userId = '${req.decoded.userId}' RETURNING *;`;
 
     pool.query(deleteEntryQuery, (err, result) => {
       // Does not show content but response is deleted.
@@ -145,6 +159,7 @@ class EntriesController {
       }
       return res.status(200).json({
         message: 'User Deleted!!',
+        userDeleted: result.rows[0],
       });
     });
 
@@ -153,13 +168,11 @@ class EntriesController {
 
   static countEntries(req, res) {
     // Get userId from token
-    const countQuery = 'SELECT COUNT(entriesId) FROM entries';
-    pool.query(countQuery, (err, result) => {
-      return res.status(200).json({
-        message: 'Entries count',
-        count: result.rows[0],
-      });
-    });
+    const countQuery = `SELECT COUNT(entriesId) FROM entries WHERE userId = '${req.decoded.userId}'`;
+    pool.query(countQuery, (err, result) => res.status(200).json({
+      message: 'Entries count',
+      count: result.rows[0],
+    }));
     return null;
   }
 }
